Added check E111 - E116
The new errors are below and tests have been added. Also
there was a panic in the whitespace checker that is now
handled.

E111: indentation is not a multiple of four
E112: expected an indented block
E113: unexpected indentation
E114: indentation is not a multiple of four (comment)
E115: expected an indented block (comment)
E116: unexpected indentation (comment)